refactor: Error 상수 파일 네이밍 변경, 에러 메세지 추가
