+++ src/js/constants/errorMessage.js
@@ -2,4 +2,5 @@ export const ERROR_MESSAGE = {
   VIDEO_MISSING_PROPERTY: '누락된 데이터가 있습니다.',
   SAVE_VIDEO_COUNT_OVER: '100개까지만 저장이 가능합니다.',
   ALREADY_SAVED_VIDEO: '이미 저장한 비디오입니다.',
+  EMPTY_KEYWORD: '키워드를 입력해주세요',
 };
